워크스루 로직 위치 변경
