+++ appFrontend/src/navigation/MainBottomTabNavigation.tsx
@@ -13,6 +13,10 @@ import IconMy from '@assets/icons/tab_my.svg';
 import { FontText } from '@/global/ui';
 import { Platform } from 'react-native';
 import { MyPageNavigation } from '.';
+import notifee from '@notifee/react-native';
+import { Walkthrough } from '@/screens/homeScreen/components';
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import { useCallback, useEffect, useState } from 'react';
 
 const Tab = createBottomTabNavigator();
 
@@ -25,75 +29,105 @@ const screenOption: BottomTabNavigationOptions = {
   },
 };
 
+export type isFirstRunType = 'isFirstRun' | 'finishedWalkThrough' | 'notFirstRun';
+
 const MainBottomTabNavigation = () => {
+  // 첫 실행이면 워크스루 띄우기
+  const [isFirstRun, setIsFirstRun] = useState<isFirstRunType>('notFirstRun');
+
+  const checkFirstRun = useCallback(async () => {
+    const hasRunBefore = await AsyncStorage.getItem('hasRunBefore');
+    if (!hasRunBefore) {
+      await AsyncStorage.setItem('hasRunBefore', 'true');
+      setIsFirstRun('isFirstRun');
+    }
+  }, []);
+
+  useEffect(() => {
+    checkFirstRun();
+  }, []);
+
+  // 워크스루 종료 시 알림 권한 요청
+  useEffect(() => {
+    if (isFirstRun === 'finishedWalkThrough') {
+      const requestPermission = async () => {
+        await notifee.requestPermission();
+      };
+      requestPermission();
+    }
+  }, [isFirstRun]);
+
   return (
-    <Tab.Navigator initialRouteName={HOME} screenOptions={screenOption}>
-      <Tab.Screen
-        name="homeStack"
-        component={HomeNavigation}
-        options={{
-          tabBarLabel: ({ focused }) => (
-            <FontText
-              text="홈"
-              className={cn('text-10 mt-5', {
-                'text-light-blue': focused,
-                'text-gray-d7d': !focused,
-              })}
-              fontWeight="600"
-            />
-          ),
-          tabBarIcon: ({ focused }) => {
-            return <>{focused ? <IconFocusedMap /> : <IconUnFocusedMap />}</>;
-          },
-        }}
-      />
-      <Tab.Screen
-        name="NowScreen"
-        component={NowScreen}
-        options={{
-          tabBarLabel: ({ focused }) => (
-            <FontText
-              text="NOW"
-              className={cn('text-10 mt-5', {
-                'text-light-blue': focused,
-                'text-gray-d7d': !focused,
-              })}
-              fontWeight="600"
-            />
-          ),
-          tabBarIcon: ({ focused }) => (
-            <IconNow
-              color={focused ? COLOR.LIGHT_BLUE : 'transparent'}
-              strokeWidth={focused ? 0 : 1}
-              stroke={COLOR.GRAY_7D}
-            />
-          ),
-        }}
-      />
-      <Tab.Screen
-        name={MY_ROOT}
-        component={MyPageNavigation}
-        options={{
-          tabBarLabel: ({ focused }) => (
-            <FontText
-              text="마이"
-              className={cn('text-10 mt-5', {
-                'text-light-blue': focused,
-                'text-gray-d7d': !focused,
-              })}
-              fontWeight="600"
-            />
-          ),
-          tabBarIcon: ({ focused }) => (
-            <IconMy
-              color={focused ? COLOR.LIGHT_BLUE : 'transparent'}
-              strokeWidth={focused ? 0 : 1}
-              stroke={COLOR.GRAY_7D}
-            />
-          ),
-        }}
-      />
-    </Tab.Navigator>
+    <>
+      <Tab.Navigator initialRouteName={HOME} screenOptions={screenOption}>
+        <Tab.Screen
+          name="homeStack"
+          component={HomeNavigation}
+          options={{
+            tabBarLabel: ({ focused }) => (
+              <FontText
+                text="홈"
+                className={cn('text-10 mt-5', {
+                  'text-light-blue': focused,
+                  'text-gray-d7d': !focused,
+                })}
+                fontWeight="600"
+              />
+            ),
+            tabBarIcon: ({ focused }) => {
+              return <>{focused ? <IconFocusedMap /> : <IconUnFocusedMap />}</>;
+            },
+          }}
+        />
+        <Tab.Screen
+          name="NowScreen"
+          component={NowScreen}
+          options={{
+            tabBarLabel: ({ focused }) => (
+              <FontText
+                text="NOW"
+                className={cn('text-10 mt-5', {
+                  'text-light-blue': focused,
+                  'text-gray-d7d': !focused,
+                })}
+                fontWeight="600"
+              />
+            ),
+            tabBarIcon: ({ focused }) => (
+              <IconNow
+                color={focused ? COLOR.LIGHT_BLUE : 'transparent'}
+                strokeWidth={focused ? 0 : 1}
+                stroke={COLOR.GRAY_7D}
+              />
+            ),
+          }}
+        />
+        <Tab.Screen
+          name={MY_ROOT}
+          component={MyPageNavigation}
+          options={{
+            tabBarLabel: ({ focused }) => (
+              <FontText
+                text="마이"
+                className={cn('text-10 mt-5', {
+                  'text-light-blue': focused,
+                  'text-gray-d7d': !focused,
+                })}
+                fontWeight="600"
+              />
+            ),
+            tabBarIcon: ({ focused }) => (
+              <IconMy
+                color={focused ? COLOR.LIGHT_BLUE : 'transparent'}
+                strokeWidth={focused ? 0 : 1}
+                stroke={COLOR.GRAY_7D}
+              />
+            ),
+          }}
+        />
+      </Tab.Navigator>
+      {isFirstRun === 'isFirstRun' && <Walkthrough setIsFirstRun={setIsFirstRun} />}
+    </>
   );
 };
